Add token validation and current user permissions

+++ src/v1/posts/postsController.js
@@ -1,5 +1,6 @@
 import * as postsDAL from './postsDAL';
 import { validate } from './Post';
+import { ForbiddenError, NotFoundError } from '../../services/errorService';
 
 export const getList = async (req, res, next) => {
   try {
@@ -26,8 +27,9 @@ export const getItem = async (req, res, next) => {
 export const createItem = async (req, res, next) => {
   try {
     const { body } = await validate(req);
+    const { id: user_id } = req.currentUser;
 
-    const result = await postsDAL.createRecord(body);
+    const result = await postsDAL.createRecord({ ...body, user_id });
     res.status(200).json({ result });
   } catch (err) {
     next(err);
@@ -37,6 +39,17 @@ export const createItem = async (req, res, next) => {
 export const updateItem = async (req, res, next) => {
   try {
     const { params, body } = await validate(req);
+    const { id: user_id } = req.currentUser;
+
+    const existRecord = await postsDAL.getRecord(params.id);
+
+    if (!existRecord) {
+      throw new NotFoundError();
+    }
+
+    if (existRecord.user_id !== user_id) {
+      throw new ForbiddenError();
+    }
 
     const result = await postsDAL.updateRecord(params.id, body);
     res.status(200).json({ result });
@@ -48,6 +61,18 @@ export const updateItem = async (req, res, next) => {
 export const deleteItem = async (req, res, next) => {
   try {
     const { params } = await validate(req);
+    const { id: user_id } = req.currentUser;
+
+    const existRecord = await postsDAL.getRecord(params.id);
+
+    if (!existRecord) {
+      throw new NotFoundError();
+    }
+
+    if (existRecord.user_id !== user_id) {
+      throw new ForbiddenError();
+    }
+
     const result = await postsDAL.deleteRecord(params.id);
 
     res.status(200).json({ result });
